pyproject.toml
[project]
name = "werewolf"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "pydantic-ai>=1.88.0",
    "pydantic>=2.7.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pyyaml>=6.0",
    "jinja2>=3.1",
]

[dependency-groups]
dev = [
    "pytest-playwright>=0.7.2",
]

[tool.pytest.ini_options]
asyncio_mode = "strict"

[tool.playwright]
browser = "chromium"
